+++ packages/ui/src/finance/DefaultFinanceCard.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+type Props = {
+    url: string;
+};
+
+const DefaultFinanceCard = ({ url }: Props) => {
+
+    return (
+        <button className="w-[65vw] h-[20vh] rounded-lg bg-gray-500">
+            <a href={url}></a>
+        </button>
+    );
+};
+
+export default DefaultFinanceCard;
+
