+++ packages/ui/src/index.tsx
@@ -9,6 +9,8 @@ import { Button } from "./button";
 import DashboardNav from "./DashboardNav";
 import DashboardCard from "./dashboard/DashboardCard";
 import PropertyList from "./properties/PropertyList";
+import PropertyComponent from "./properties/PropertyComponent";
+import DefaultFinanceCard from "./finance/DefaultFinanceCard";
 
 export {
     Navbar,
@@ -22,4 +24,6 @@ export {
     DashboardNav,
     DashboardCard,
     PropertyList,
+    PropertyComponent,
+    DefaultFinanceCard,
 };
